ad_view modified

+++ src/by/vasiliuk/project/service/AdvertService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface AdvertService {
      List<Advert> findAllAds() throws ServiceException;
-     Optional<Advert> findAdById(long id) throws ServiceException;
+     Optional<Advert> findAdById(int id) throws ServiceException;
      List<Advert> findBySectionId(long id) throws ServiceException;
      void saveAdvert(String title, String text, long id, String section) throws ServiceException;
      List<Advert> editAdvert(String advertTitle, String advertText) throws ServiceException;
